Attempt cast for duplicate names with differing types
On several occasions the newer nif.xml now has version-exclusive names which are identical yet have different types.  Such as uint in one version and ushort in another.  Adding the cast to read/write will adjust the read/write size appropriately.

Note this does not fix all duplicate names, just the ones where a cast during read/write is all that is necessary.

The first type also has to be the largest as declaration for the type depends on the first name, and subsequent smaller types will reuse that name.